+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added passing `collectLog` function to `Genemo.reproduce` as the 3rd argument.
 - Changed `Genemo.logIterationData` API.
 - Renamed `debugData` key to `logs` in `iterationCallback`.
+- Added `collectLogs` param to the runner function (`Genemo.run`).
 
 ## 0.7.0 - 2019-07-18
 - Added `maxBlockingTime` option to `Genemo.runEvolutionAsync`.